LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/inline/getAll-shorthand-expected.txt


PASS StylePropertyMap.getAll() with a shorthand property set explicitly in inline style returns a base CSSStyleValue
PASS StylePropertyMap.getAll() with a shorthand property that is partially in inline style returns empty list
